+++ exch/exmdb/main.cpp
@@ -149,6 +149,9 @@ BOOL SVC_exmdb_provider(enum plugin_op reason, const struct dlfuncs &ppdata)
 		}
 		if (!exmdb_provider_reload(nullptr, pconfig))
 			return false;
+		g_exmdb_client_only = strcasecmp(get_prog_id(), "istore") != 0;
+		if (g_exmdb_client_only)
+			return TRUE;
 
 		auto listen_ip = pconfig->get_value("listen_ip");
 		uint16_t listen_port = pconfig->get_ll("exmdb_listen_port");
@@ -225,7 +228,7 @@ BOOL SVC_exmdb_provider(enum plugin_op reason, const struct dlfuncs &ppdata)
 			db_engine_stop();
 			return FALSE;
 		}
-		if (0 != exmdb_listener_trigger_accept()) {
+		if (!g_exmdb_client_only && exmdb_listener_trigger_accept() != 0) {
 			mlog(LV_ERR, "exmdb_provider: failed to start exmdb listener");
 			exmdb_listener_stop();
 			exmdb_parser_stop();